discord-bot/requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
